move readers and detectors into the Inference folder
