+++ SerialPrograms/Source/NintendoSwitch/DevPrograms/TestProgramSwitch.cpp
@@ -123,8 +123,8 @@
 #include "PokemonSV/Programs/AutoStory/PokemonSV_AutoStory.h"
 #include "PokemonSV/Programs/AutoStory/PokemonSV_MenuOption.h"
 #include "PokemonSV/Programs/AutoStory/PokemonSV_MenuOptionDatabase.h"
-#include "PokemonSV/Programs/AutoStory/PokemonSV_TutorialDetector.h"
-#include "PokemonSV/Programs/AutoStory/PokemonSV_PokemonMovesReader.h"
+#include "PokemonSV/Inference/PokemonSV_TutorialDetector.h"
+#include "PokemonSV/Inference/PokemonSV_PokemonMovesReader.h"
 
 
 #include <QPixmap>
