--- SerialPrograms/Source/PokemonSV/Programs/AutoStory/PokemonSV_PokemonMovesReader.h
@@ -1,63 +1,0 @@
-/*  Pokemon Moves Reader
- *
- *  From: https://github.com/PokemonAutomation/Arduino-Source
- *
- */
-
-#ifndef PokemonAutomation_PokemonSV_PokemonMovesReader_H
-#define PokemonAutomation_PokemonSV_PokemonMovesReader_H
-
-#include <map>
-#include <array>
-#include "Common/Cpp/Containers/FixedLimitVector.h"
-#include "CommonFramework/Language.h"
-#include "CommonFramework/VideoPipeline/VideoOverlayScopes.h"
-#include "CommonFramework/OCR/OCR_SmallDictionaryMatcher.h"
-
-namespace PokemonAutomation{
-    class AsyncDispatcher;
-    class ConsoleHandle;
-    class BotBaseContext;
-namespace NintendoSwitch{
-namespace PokemonSV{
-
-class PokemonMovesOCR : public OCR::SmallDictionaryMatcher{
-    static constexpr double MAX_LOG10P = -1.30;
-    static constexpr double MAX_LOG10P_SPREAD = 0.50;
-
-public:
-    PokemonMovesOCR();
-
-    static PokemonMovesOCR& instance();
-
-    OCR::StringMatchResult read_substring(
-        Logger& logger,
-        Language language,
-        const ImageViewRGB32& image,
-        const std::vector<OCR::TextColorRange>& text_color_ranges,
-        double min_text_ratio = 0.01, double max_text_ratio = 0.50
-    ) const;
-
-    
-};    
-
-class PokemonMovesReader{
-
-public:
-    PokemonMovesReader(Language language);
-
-    void make_overlays(VideoOverlaySet& items) const;
-
-    std::string read_move(Logger& logger, const ImageViewRGB32& screen, int8_t index) const;
-
-private:
-    Language m_language;
-    std::array<ImageFloatBox, 4> m_boxes_rearrange;
-};
-
-
-
-}
-}
-}
-#endif
